use only 16 bits of timers
even if the timer is 32 bits, use only bottom 16 bits for consistency

+++ Mcu/v203/Inc/blutil.h
@@ -116,7 +116,7 @@ static inline void bl_timer_disable(void)
     RCC_APB2PeriphClockCmd(RCC_APB2Periph_TIM1, DISABLE);
 }
 
-static inline uint32_t bl_timer_us(void)
+static inline uint16_t bl_timer_us(void)
 {
     return BL_TIMER->CNT;
 }
